+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed diagnostics not clearing for files when workspace diagnostics is not enabled
 - Fixed metatable name not being used when hovering over the function of a metatable
 - Manually increased some internal limits to reduce likelihood of type errors
+- Fixed diagnostics (and other global configuration) not loading when not inside of a workspace
 
 ## [1.10.1] - 2022-09-24